+++ braindecode/models/__init__.py
@@ -16,7 +16,7 @@ from .eegresnet import EEGResNet
 from .eeginception_erp import EEGInceptionERP
 from .eeginception_mi import EEGInceptionMI
 from .atcnet import ATCNet
-from .tcn import TCN
+from .tcn import TCN, BDTCN
 from .sleep_stager_chambon_2018 import SleepStagerChambon2018
 from .sleep_stager_blanco_2020 import SleepStagerBlanco2020
 from .sleep_stager_eldele_2021 import SleepStagerEldele2021